FE:[refactor] map marker 성능 향상 #65
